pyproject.toml
[project]
name = "recruit-calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-api-python-client>=2.167.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "lxml>=5.4.0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "selenium>=4.31.0",
]
